Move code to single part
