+++ src/Workspaces/Remote/Core/SolutionAssetStorage.Scope.cs
@@ -48,5 +48,122 @@ internal partial class SolutionAssetStorage
 
         public void Dispose()
             => _storage.DecreaseScopeRefCount(this);
+
+        /// <summary>
+        /// Retrieve asset of a specified <paramref name="checksum"/> available within <see langword="this"/> from
+        /// the storage.
+        /// </summary>
+        public async ValueTask<SolutionAsset?> GetAssetAsync(Checksum checksum, CancellationToken cancellationToken)
+        {
+            if (checksum == Checksum.Null)
+            {
+                // check nil case
+                return SolutionAsset.Null;
+            }
+
+            var remotableData = await FindAssetAsync(checksum, cancellationToken).ConfigureAwait(false);
+            if (remotableData != null)
+            {
+                return remotableData;
+            }
+
+            // if it reached here, it means things get canceled. due to involving 2 processes,
+            // current design can make slightly staled requests to running even when things canceled.
+            // if it is other case, remote host side will throw and close connection which will cause
+            // vs to crash.
+            // this should be changed once I address this design issue
+            cancellationToken.ThrowIfCancellationRequested();
+
+            return null;
+        }
+
+        /// <summary>
+        /// Retrieve assets of specified <paramref name="checksums"/> available within <see langword="this"/> from
+        /// the storage.
+        /// </summary>
+        public async ValueTask<IReadOnlyDictionary<Checksum, SolutionAsset>> GetAssetsAsync(
+            IEnumerable<Checksum> checksums, CancellationToken cancellationToken)
+        {
+            using var checksumsToFind = Creator.CreateChecksumSet(checksums);
+
+            var numberOfChecksumsToSearch = checksumsToFind.Object.Count;
+            var result = new Dictionary<Checksum, SolutionAsset>(numberOfChecksumsToSearch);
+
+            if (checksumsToFind.Object.Remove(Checksum.Null))
+            {
+                result[Checksum.Null] = SolutionAsset.Null;
+            }
+
+            await FindAssetsAsync(checksumsToFind.Object, result, cancellationToken).ConfigureAwait(false);
+            if (result.Count == numberOfChecksumsToSearch)
+            {
+                // no checksum left to find
+                Debug.Assert(checksumsToFind.Object.Count == 0);
+                return result;
+            }
+
+            // if it reached here, it means things get canceled. due to involving 2 processes,
+            // current design can make slightly staled requests to running even when things canceled.
+            // if it is other case, remote host side will throw and close connection which will cause
+            // vs to crash.
+            // this should be changed once I address this design issue
+            cancellationToken.ThrowIfCancellationRequested();
+
+            return result;
+        }
+
+        /// <summary>
+        /// Find an asset of the specified <paramref name="checksum"/> within <see langword="this"/>.
+        /// </summary>
+        private async ValueTask<SolutionAsset?> FindAssetAsync(Checksum checksum, CancellationToken cancellationToken)
+        {
+            using var checksumPool = Creator.CreateChecksumSet(SpecializedCollections.SingletonEnumerable(checksum));
+            using var resultPool = Creator.CreateResultSet();
+
+            await FindAssetsAsync(checksumPool.Object, resultPool.Object, cancellationToken).ConfigureAwait(false);
+
+            if (resultPool.Object.Count == 1)
+            {
+                var (resultingChecksum, value) = resultPool.Object.First();
+                Contract.ThrowIfFalse(checksum == resultingChecksum);
+
+                return new SolutionAsset(checksum, value);
+            }
+
+            return null;
+        }
+
+        /// <summary>
+        /// Find an assets of the specified <paramref name="remainingChecksumsToFind"/> within <see
+        /// langword="this"/>. Once an asset of given checksum is found the corresponding asset is placed to
+        /// <paramref name="result"/> and the checksum is removed from <paramref name="remainingChecksumsToFind"/>.
+        /// </summary>
+        private async Task FindAssetsAsync(HashSet<Checksum> remainingChecksumsToFind, Dictionary<Checksum, SolutionAsset> result, CancellationToken cancellationToken)
+        {
+            using var resultPool = Creator.CreateResultSet();
+
+            await FindAssetsAsync(remainingChecksumsToFind, resultPool.Object, cancellationToken).ConfigureAwait(false);
+
+            foreach (var (checksum, value) in resultPool.Object)
+            {
+                result[checksum] = new SolutionAsset(checksum, value);
+            }
+        }
+
+        private async Task FindAssetsAsync(HashSet<Checksum> remainingChecksumsToFind, Dictionary<Checksum, object> result, CancellationToken cancellationToken)
+        {
+            var solutionState = this.Solution;
+            if (solutionState.TryGetStateChecksums(out var stateChecksums))
+                await stateChecksums.FindAsync(solutionState, remainingChecksumsToFind, result, cancellationToken).ConfigureAwait(false);
+
+            foreach (var projectId in solutionState.ProjectIds)
+            {
+                if (remainingChecksumsToFind.Count == 0)
+                    break;
+
+                if (solutionState.TryGetStateChecksums(projectId, out var checksums))
+                    await checksums.FindAsync(solutionState, remainingChecksumsToFind, result, cancellationToken).ConfigureAwait(false);
+            }
+        }
     }
 }
